Some suggestions from The Wordsmith
Just a few grammatical tweaks. The only one I'm not sure of is on line 141, where I change 'a function which accepts one argument of component to create a HOC' into 'a function which accepts one argument (ie, 'component') to create a HOC': as I may have misunderstood the meaning.